+++ packages/sapa/src/App.js
@@ -92,13 +92,6 @@ export const hydrate = (ElementClass, opt = {}) => {
 export const potal = (ElementClass, opt = {}) => {
   const $container = Dom.create(opt.container || document.body);
 
-  // component 일 때는 바로 렌더링을 시작한다.
-  if (ElementClass instanceof VNodeComponent) {
-    DomRenderer(ElementClass, $container);
-
-    return ElementClass.instance;
-  }
-
   if (ElementClass instanceof VNode) {
     const rootVNode = ElementClass;
     ElementClass = () => rootVNode;