+++ packages/ui/src/components/property-editor/editor-items/ColorItem.jsx
@@ -5,7 +5,10 @@ export function ColorItem({ value, onChange, item }) {
   return (
     <InputPaint
       value={value}
-      onChange={onChange}
+      sync
+      onChange={(color, inputPaintInstance) => {
+        onChange && onChange(color, inputPaintInstance);
+      }}
       onClickColorView={(e, color) => {
         onClickColorView && onClickColorView(e, color);
       }}